Исрпавлены некоторые ошибки с контекстом
