+++ ViplataProject.Infrastructure/App.config
@@ -19,6 +19,6 @@
     </DbProviderFactories>
   </system.data>
   <connectionStrings>
-    <add name="Contxt" connectionString="data source=C:\ViplataProject\Viplata_project2_db.db.sql" providerName="System.Data.SQLite.EF6" />
+    <add name="Context" connectionString="data source=C:\ViplataProject\VP.db" providerName="System.Data.SQLite.EF6" />
   </connectionStrings>
 </configuration>
