+++ ViplataProject.Infrastructure/Database/EmployeeRepository.cs
@@ -12,7 +12,7 @@ namespace ViplataProject.Infrastructure.Database
     {
         public List<EmployeeViewModel> GetList()
         {
-            using (var context = new Contxt())
+            using (var context = new Context())
             {
                 var items = context.Employee.ToList();
                 return EmployeeMapper.Map(items);
@@ -20,7 +20,7 @@ namespace ViplataProject.Infrastructure.Database
         }
         public EmployeeViewModel GetById(long id)
         {
-            using (var context = new Contxt())
+            using (var context = new Context())
             {
                 var item = context.Employee.FirstOrDefault(x => x.ID == id);
                 return EmployeeMapper.Map(item);
