+++ ViplataProject.Infrastructure/JobEntity.cs
