+++ ViplataProject.Infrastructure/Mappers/EmployeeMapper.cs
@@ -19,8 +19,8 @@ namespace ViplataProject.Infrastructure.Mappers
                 Patronymic = entity.Patronymic,
                 Date_Of_Birth = entity.Date_Of_Birth,
                 Gender = entity.Gender == "м"? "мужской" : "женский",
-                Disability = entity.Disability,
-                Expirience = entity.Expirience,
+                disability = entity.disability,
+                expirience = entity.expirience,
                 UserId = entity.UserId,
                 JobId = entity.JobId,
                 CompanyId = entity.CompanyId,
