+++ ViplataProject.Infrastructure/RoleEntity.cs
