+++ ViplataProject.Infrastructure/UserEntity.cs
