+++ ViplataProject.Infrastructure/ViewModels/EmployeeViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,8 +15,8 @@ namespace ViplataProject.Infrastructure.ViewModels
         public string Patronymic { get; set; }
         public string Date_Of_Birth { get; set; }
         public string Gender { get; set; }
-        public string Disability { get; set; }
-        public string Expirience { get; set; }
+        public string disability { get; set; }
+        public string expirience { get; set; }
         public long UserId { get; set; }
         public long JobId { get; set; }
         public long CompanyId { get; set; }
