+++ ViplataProject.Infrastructure/ViplataProject.Infrastructure.csproj
@@ -65,7 +65,7 @@
   <ItemGroup>
     <Compile Include="CompanyEntity.cs" />
     <Compile Include="Consts\UserInfoConsts.cs" />
-    <Compile Include="Contxt.cs" />
+    <Compile Include="Context.cs" />
     <Compile Include="Database\CompanyRepository.cs" />
     <Compile Include="Database\EmployeeRepository.cs" />
     <Compile Include="Database\IBaseRepository.cs" />
