+++ ViplataProject/Pages/CompaniesPage.xaml
@@ -14,7 +14,7 @@
             <Label Content="Компании"/>
             <Button Content="Выйти в меню" Click="MenuButton_Click"/>
         </StackPanel>
-        <DataGrid x:Name="CompanyDG" CanUserReorderColumns="False" IsReadOnly ="True" AutoGenerateColumns="False" ItemsSource="{Binding}"  Grid.Row="1" Grid.Column="0" Grid.RowSpan="2" Grid.ColumnSpan="2">
+        <DataGrid x:Name="CompanyDG" CanUserReorderColumns="False" CanUserSortColumns="True" IsReadOnly ="True" AutoGenerateColumns="False" ItemsSource="{Binding}"  Grid.Row="1" Grid.Column="0" Grid.RowSpan="2" Grid.ColumnSpan="2">
             <DataGrid.Columns>
                 <DataGridTextColumn Header="Название компании" Binding="{Binding Name_Company}"/>
                 <DataGridTextColumn Header="Юридический адрес" Binding="{Binding Legal_Address}"/>
